updated mnist notebook

+++ src/diffusion_3d/chestct/diffusion/unet/nn.py
@@ -302,7 +302,7 @@ class Diffusion3D(nn.Module):
             self.decoder = add_stochastic_depth_dropout(self.decoder, model_config.survival_prob)
 
     def forward(self, x, timesteps, spacings, guidance):
-        time_emb = get_timestep_embeddings_1d(self.model_config.time_channels, timesteps)
+        time_emb = get_timestep_embeddings_1d(self.model_config.time_channels, timesteps).half()
 
         x = self.encoder_mapping(x)
         encodings = self.encoder(x, time_emb, spacings, guidance)
